requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# AI/ML Dependencies
openai==1.54.0
openai-whisper==20240930
langchain==0.3.7
langchain-openai==0.2.8
langchain-google-genai==2.0.5
langchain-core==0.3.15
langgraph==0.2.45

# Document Generation
python-docx==1.1.2

# Audio Processing
ffmpeg-python==0.2.0

# Data Validation
pydantic==2.9.2
